Fixed tool functionaility + implemented SaveOutputToTextFile_v2
In the Tauri config, for the gobuster command, the ‘G’ was capitalised which is the cause of the issue, since the command itself should be executed in all lower case. This has been modified and tested successfully.

A bug I found with this tool, the existing Output file textbox didn't do anything, and no Output file was created upon running the tool. 

I've modified the code to remove the existing Output textbox, and I've instead implemented the SaveOutputToTextFile_v2 function as per the documentation in Teams. After modifying the code and implementing the SaveOutputToTextFile_v2 function all is working as intended and the output file will be created.