Update DNSMap.tsx
Updated DNSmap per the App Attack report as well as by personal findings
• Corrected spelling errors in tool description
• Corrected arguments called for advanced features to be in line with the expected CLI input
External wordlist: -S to -w
Output to CSV: -s to -c
IP to ignore: -t to -i

Additionally corrected the functions for calling the advanced mode features. Previously they were causing error messages due to causing issues with the syntax of the command when unused

One persisting small error: the "Save to CSV" feature is saving the files into the /src-tauri/ folder within the DDT instead of the /OutputFiles/ folder. Yet to fix this