Update arpaname.tsx
The "dependencies" variable is designed to hold the command that the respective tool uses, so the DDT can check if the command is installed on the machine. The value of this variable was previously incorrect, causing the DDT to think that the arpane tool wasn't installed, regardless of if it was installed or not. This change amends the value of this variable so that the DDT can correctly determine if the arpaname tool is installed on the machine.