Update Rainbowcrack.tsx
The dependencies variable contains the command associated with a given tool. If this value is incorrect for a specific tool, then the DDT will be unable to tell if the tool is installed on a system, preventing users from accessing the tool. This change has adjusted the value of the dependencies variable for the rainbowcrack tool so that the DDT can correctly determine if the tool is installed on the system.