Issue #692 Fix (#825)
* Update arpaname.tsx

The "dependencies" variable is designed to hold the command that the respective tool uses, so the DDT can check if the command is installed on the machine. The value of this variable was previously incorrect, causing the DDT to think that the arpane tool wasn't installed, regardless of if it was installed or not. This change amends the value of this variable so that the DDT can correctly determine if the arpaname tool is installed on the machine.

* Update Rainbowcrack.tsx

The dependencies variable contains the command associated with a given tool. If this value is incorrect for a specific tool, then the DDT will be unable to tell if the tool is installed on a system, preventing users from accessing the tool. This change has adjusted the value of the dependencies variable for the rainbowcrack tool so that the DDT can correctly determine if the tool is installed on the system.