Update References.tsx
Despite running npx prettier -write, the previous version of the file didn't pass the CI check. I ran the command again and used the diff command with the second version of the file and found that the only difference is some spaces on line 381. Hopefully the file should pass the CI check after this.