Update References.tsx (#950)
* Update References.tsx

* Update References.tsx

Ran "npx prettier --write src/pages/References.tsx" and uploaded the changes to make sure it passes the CI check.

* Update References.tsx

Despite running npx prettier -write, the previous version of the file didn't pass the CI check. I ran the command again and used the diff command with the second version of the file and found that the only difference is some spaces on line 381. Hopefully the file should pass the CI check after this.

* Update References.tsx

Added full stops to the descriptions, arranged the added tools into alphabetical order and placed CVE-2022-27925 in the correct place.